tenth/requirements.txt
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2020.6.20
chardet==3.0.4
Django==2.2.5
django-ckeditor==5.9.0
django-cors-headers==3.4.0
django-haystack==2.8.1
django-js-asset==1.2.2
django-redis==4.12.1
django-rest-framework-mongoengine==3.4.1
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
drf-haystack==1.8.7
elasticsearch==2.4.1
idna==2.10
mongoengine==0.20.0
mutagen==1.44.0
mysqlclient==2.0.1
Pillow==7.2.0
PyJWT==1.7.1
pymongo==3.11.0
python-dateutil==2.8.1
pytz==2020.1
redis==3.5.3
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
urllib3==1.25.10
